requirements.txt
Django~=5.2.1
djangorestframework~=3.14.0
psycopg2-binary~=2.9.9
pytest~=8.0.0
pytest-django~=4.8.0
pandas~=2.2.0
requests~=2.31.0
avro-python3~=1.10.2